requirements.txt
google-adk
google-cloud-aiplatform
google-generativeai
python-multipart
python-dotenv